Add Segment.new/1
